git init

+++ src/main/java/com/stephen/todo/TodoLogicImplement.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TodoLogicImplement implements TodoLogic{
@@ -28,26 +29,59 @@ public class TodoLogicImplement implements TodoLogic{
     }
 
     @Override
-    public List<Todo> getAllByDate(){
+    public List<Todo> getAllByDate(Integer userId){
         List<Todo> todos = new ArrayList<>();
-        todoRepository.findAllByOrderByDateAsc().forEach(todos::add);
+        todoRepository.findByUserIdOrderByDateAsc(userId)
+                .forEach(todos::add);
+
+            return todos;
+    }
+
+    @Override
+    public List<Todo> getAllByName(Integer userId){
+        List<Todo> todos = new ArrayList<>();
+        todoRepository.findByUserIdOrderByNameAsc(userId)
+                .forEach(todos::add);
 
         return todos;
     }
 
     @Override
-    public Todo updateTodo(Integer id, Todo todo) {
-        Todo todo1 = todoRepository.findById(id).get();
-        todo1.setName(todo.getName());
-        todo1.setDescription(todo.getDescription());
-        todo1.setDate(todo.getDate());
-        todo1.setUser(todo.getUser());
-        return todoRepository.save(todo1);
+    public Todo updateTodo(Integer userId, Integer id, Integer vid, Todo todo) {
+        if(!todoRepository.findByUserIdAndId(userId, id).isEmpty()) {
+            todo = todoRepository.findById(id).get();
+            return todoRepository.save(todo);
+        }
+        return null;
+//            Todo todo1 = todoRepository.findById(id).get();
+//            if(!todo.getName().isEmpty()) {
+//                todo1.setName(todo.getName());
+//            }
+//            if(!todo.getDescription().isEmpty()) {
+//                todo1.setDescription(todo.getDescription());
+//            }
+//            todo1.setDate(todo.getDate());
+//            todo1.setUser(todo.getUser());
+//            todoRepository.save(todo1);
+//            todo = todoRepository.findById(id).get();
+
+//            todo = todoRepository.findById(id).get();
+//            return todoRepository.save(todo);
+//            return true;
+//        }else{
+//            return false;
+//        }
     }
 
     @Override
-    public void deleteTodo(Integer id) {
-        todoRepository.deleteById(id);
+    public Boolean deleteTodo(Integer userId, Integer id) {
+        if(!todoRepository.findByUserIdAndId(userId, id).isEmpty()){
+            todoRepository.deleteById(id);
+            return true;
+        }else{
+            return false;
+        }
+
     }
